Convert depths from hyperspectral imaging

Convert depths from hyperspectral imaging

hsi_to_dblf(corename, cm, extraAllowedBottom = 1)

Arguments

corename

a vector of corenames

cm

a vector of corresponding depths that matches the length of corenames

extraAllowedBottom

how many cm below the bottom of the ROI are allowed (converted to bottom of ROI) (default = 1)

Value

depth below lake floor in cm

Examples

hsi_to_dblf("L380_DUNCA3_LC4U_2",0:30)
#>              corename coreSectionDepth  dblf topSource botSource
#> 1  L380_DUNCA3_LC4U_2                0  89.5   non-HSI   non-HSI
#> 2  L380_DUNCA3_LC4U_2                1  90.5   non-HSI   non-HSI
#> 3  L380_DUNCA3_LC4U_2                2  91.5   non-HSI   non-HSI
#> 4  L380_DUNCA3_LC4U_2                3  92.5   non-HSI   non-HSI
#> 5  L380_DUNCA3_LC4U_2                4  93.5   non-HSI   non-HSI
#> 6  L380_DUNCA3_LC4U_2                5  94.5   non-HSI   non-HSI
#> 7  L380_DUNCA3_LC4U_2                6  95.5   non-HSI   non-HSI
#> 8  L380_DUNCA3_LC4U_2                7  96.5   non-HSI   non-HSI
#> 9  L380_DUNCA3_LC4U_2                8  97.5   non-HSI   non-HSI
#> 10 L380_DUNCA3_LC4U_2                9  98.5   non-HSI   non-HSI
#> 11 L380_DUNCA3_LC4U_2               10  99.5   non-HSI   non-HSI
#> 12 L380_DUNCA3_LC4U_2               11 100.5   non-HSI   non-HSI
#> 13 L380_DUNCA3_LC4U_2               12 101.5   non-HSI   non-HSI
#> 14 L380_DUNCA3_LC4U_2               13 102.5   non-HSI   non-HSI
#> 15 L380_DUNCA3_LC4U_2               14 103.5   non-HSI   non-HSI
#> 16 L380_DUNCA3_LC4U_2               15 104.5   non-HSI   non-HSI
#> 17 L380_DUNCA3_LC4U_2               16 105.5   non-HSI   non-HSI
#> 18 L380_DUNCA3_LC4U_2               17 106.5   non-HSI   non-HSI
#> 19 L380_DUNCA3_LC4U_2               18 107.5   non-HSI   non-HSI
#> 20 L380_DUNCA3_LC4U_2               19 108.5   non-HSI   non-HSI
#> 21 L380_DUNCA3_LC4U_2               20 109.5   non-HSI   non-HSI
#> 22 L380_DUNCA3_LC4U_2               21 110.5   non-HSI   non-HSI
#> 23 L380_DUNCA3_LC4U_2               22 111.5   non-HSI   non-HSI
#> 24 L380_DUNCA3_LC4U_2               23 112.5   non-HSI   non-HSI
#> 25 L380_DUNCA3_LC4U_2               24 113.5   non-HSI   non-HSI
#> 26 L380_DUNCA3_LC4U_2               25 114.5   non-HSI   non-HSI
#> 27 L380_DUNCA3_LC4U_2               26 115.5   non-HSI   non-HSI
#> 28 L380_DUNCA3_LC4U_2               27 116.5   non-HSI   non-HSI
#> 29 L380_DUNCA3_LC4U_2               28 117.5   non-HSI   non-HSI
#> 30 L380_DUNCA3_LC4U_2               29 118.5   non-HSI   non-HSI
#> 31 L380_DUNCA3_LC4U_2               30 119.5   non-HSI   non-HSI
#>              coreName compactionAdjusted compositeSequence
#> 1  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 2  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 3  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 4  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 5  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 6  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 7  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 8  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 9  L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 10 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 11 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 12 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 13 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 14 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 15 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 16 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 17 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 18 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 19 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 20 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 21 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 22 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 23 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 24 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 25 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 26 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 27 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 28 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 29 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 30 L380_DUNCA3_LC4U_2              FALSE             FALSE
#> 31 L380_DUNCA3_LC4U_2              FALSE             FALSE